perf(web): code-split heavy dashboard sub-features; lower lazy-chunk budget (#2782)
Before: route-dashboard lazy chunk was 342.12 KiB gzip with the temporary 400 KiB lazy budget.

After: route-dashboard is 209,506 bytes (210.91 KiB) gzip; max lazy chunk is route-ledger at 213,501 bytes, so maxLazyChunkGzipBytes is lowered to 215,000.

Lazy-loaded dashboard AI query, coach, things-to-check scam alerts, tax reserve guidance, warranty dashboard, mood journal, and customize panel while keeping the dashboard shell eager.

Verified: npm run build; node ../../tools/check-web-performance-budget.mjs --dist dist --budget performance.budget.json; npx vitest run src/pages/DashboardPage.test.tsx src/lib/perf --reporter=dot; npx tsc --noEmit -p tsconfig.json; npx eslint src/ --ext .ts,.tsx --max-warnings 0; npx prettier --check changed files; Playwright chromium smoke on vite preview port 4180.
